added and deleted stuff
